blip_master/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-datetimepicker==3.14
django-extensions==2.1.5
django-forms-bootstrap==3.1.0
idna==2.8
mysqlclient==1.4.1
Pillow==5.4.1
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
